+++ inst/doc/intro_vignette.R
@@ -70,21 +70,37 @@ knitr::kable(head(example_pathfindR_output, 2))
 #    top_terms = 15
 #  )
 
-## ----KEGG_vis, eval=FALSE-----------------------------------------------------
+## ----KEGG_vis-----------------------------------------------------------------
 #  input_processed <- input_processing(example_pathfindR_input)
-#  visualize_terms(
+#  gg_list <- visualize_terms(
 #    result_df = example_pathfindR_output,
 #    input_processed = input_processed,
-#    hsa_KEGG = TRUE
+#    is_KEGG_result = TRUE
+#  )  # this function returns a list of ggraph objects (named by Term ID)
+#  
+#  # save one of the plots as PDF image
+#  ggplot2::ggsave(
+#    "hsa04911_diagram.pdf",   # path to output, format is determined by extension
+#    gg_list$hsa04911,         # what to plot
+#    width = 5,                 # adjust width
+#    height = 5                # adjust height
 #  )
 
-## ----nonKEGG_viss, eval=FALSE-------------------------------------------------
+## ----nonKEGG_viss-------------------------------------------------------------
 #  input_processed <- input_processing(example_pathfindR_input)
-#  visualize_terms(
+#  gg_list <- visualize_terms(
 #    result_df = example_pathfindR_output,
 #    input_processed = input_processed,
-#    hsa_KEGG = FALSE,
+#    is_KEGG_result = FALSE,
 #    pin_name_path = "Biogrid"
+#  )  # this function returns a list of ggraph objects (named by Term ID)
+#  
+#  # save one of the plots as PDF image
+#  ggplot2::ggsave(
+#    "diabetic_cardiomyopathy_interactions.pdf",   # path to output, format is determined by extension
+#    gg_list$hsa04911,                             # what to plot
+#    width = 10,                                    # adjust width
+#    height = 6                                    # adjust height
 #  )
 
 ## ----hierarchical0------------------------------------------------------------
@@ -104,40 +120,31 @@ enrichment_chart(selected_clusters, plot_by_cluster = TRUE)
 ## ----fuzzy--------------------------------------------------------------------
 #  clustered_fuzzy <- cluster_enriched_terms(example_pathfindR_output, method = "fuzzy")
 
-## ----hmap1--------------------------------------------------------------------
-#  term_gene_heatmap(result_df = example_pathfindR_output, genes_df = example_pathfindR_input)
-
-## ----term_gene_graph----------------------------------------------------------
-#  term_gene_graph(result_df = example_pathfindR_output, use_description = TRUE)
-
-## ----upset--------------------------------------------------------------------
-#  UpSet_plot(result_df = example_pathfindR_output, genes_df = example_pathfindR_input)
-
-## ----scores, eval=TRUE, fig.height=4, fig.width=8-----------------------------
-## Vector of "Case" IDs
-cases <- c(
-  "GSM389703", "GSM389704", "GSM389706", "GSM389708",
-  "GSM389711", "GSM389714", "GSM389716", "GSM389717",
-  "GSM389719", "GSM389721", "GSM389722", "GSM389724",
-  "GSM389726", "GSM389727", "GSM389730", "GSM389731",
-  "GSM389733", "GSM389735"
-)
-
-## Calculate scores for representative terms
-## and plot heat map using term descriptions
-representative_df <- example_pathfindR_output_clustered[example_pathfindR_output_clustered$Status == "Representative", ]
-score_matrix <- score_terms(
-  enrichment_table = representative_df,
-  exp_mat = example_experiment_matrix,
-  cases = cases,
-  use_description = TRUE, # default FALSE
-  label_samples = FALSE, # default = TRUE
-  case_title = "RA", # default = "Case"
-  control_title = "Healthy", # default = "Control"
-  low = "#f7797d", # default = "green"
-  mid = "#fffde4", # default = "black"
-  high = "#1f4037" # default = "red"
-)
+## ----scores-------------------------------------------------------------------
+#  ## Vector of "Case" IDs
+#  cases <- c(
+#    "GSM389703", "GSM389704", "GSM389706", "GSM389708",
+#    "GSM389711", "GSM389714", "GSM389716", "GSM389717",
+#    "GSM389719", "GSM389721", "GSM389722", "GSM389724",
+#    "GSM389726", "GSM389727", "GSM389730", "GSM389731",
+#    "GSM389733", "GSM389735"
+#  )
+#  
+#  ## Calculate scores for representative terms
+#  ## and plot heat map using term descriptions
+#  representative_df <- example_pathfindR_output_clustered[example_pathfindR_output_clustered$Status == "Representative", ]
+#  score_matrix <- score_terms(
+#    enrichment_table = representative_df,
+#    exp_mat = example_experiment_matrix,
+#    cases = cases,
+#    use_description = TRUE, # default FALSE
+#    label_samples = FALSE, # default = TRUE
+#    case_title = "RA", # default = "Case"
+#    control_title = "Healthy", # default = "Control"
+#    low = "#f7797d", # default = "green"
+#    mid = "#fffde4", # default = "black"
+#    high = "#1f4037" # default = "red"
+#  )
 
 ## ----compare2res, eval=TRUE, fig.height=4, fig.width=8------------------------
 combined_df <- combine_pathfindR_results(
@@ -187,8 +194,8 @@ knitr::kable(head(example_pathfindR_input))
 #    gene_sets = "Custom",
 #    custom_genes = custom_genes,
 #    custom_descriptions = custom_descriptions,
-#    max_gset_size = Inf, # DO NOT LIMIT GENE SET SIZE
-#    output_dir = "misc/CREB_MYC"
+#    min_gset_size = 1,   # do not limit the gene set size for demo
+#    max_gset_size = Inf, # do not limit the gene set size for demo
 #  )
 #  
 #  knitr::kable(example_custom_genesets_result)